+++ double_click/models.py
@@ -2,97 +2,67 @@ import json
 import math
 import os
 from datetime import datetime as dt, timedelta
+from pathlib import Path
 
 from double_click.request import GeneralSession, UserSession
+
+
+class ModelAuth:
+
+    def __init__(self, requires: list = None, match_all: bool = False, **kwargs):
+        self.requires = requires
+        self.match_all = match_all
+        self.kwargs = kwargs
     
 
 class Model:
     # This could be a double_click.UserSession or double_click.GeneralSession object
-    session: GeneralSession = None  # Define a class that inherits from Model to set a default session type
-    url: str = None
+    _session: GeneralSession = None  # Define a class that inherits from Model to set a default session type
+    _url: str = None
+    _ttl = 120
+    _cache_key = None
+    _obj_identifier: any
+    _auth: ModelAuth = ModelAuth(None, False)
 
     def __init__(self, **kwargs):
-        self.url = kwargs.pop('url', self.url)
-        self.session = kwargs.pop('session', self.session)
-        if not isinstance(self.session, UserSession) and not isinstance(self.session, GeneralSession):
-            self.session = GeneralSession(disable_progress_bar=False)
+        self._url = kwargs.pop('url', self._url)
+        self._session = kwargs.pop('session', self._session)
+        if not isinstance(self._session, UserSession) and not isinstance(self._session, GeneralSession):
+            self._session = GeneralSession(disable_progress_bar=False)
 
         self._as_dict = kwargs
         for key, value in kwargs.items():
             setattr(self, key, value)
 
     @property
-    def ttl(self):
-        """Time to live. The local file cache will be refreshed after this number, in minutes, has passed.
-        """
-        return 120
-
-    @property
-    def file_path(self):
-        """Location where the http response is stored locally. If none, response will not be stored.
-        """
-        return None
-
-    @property
-    def key_identifier(self):
-        """The dictionary key to use for grouping the objects by, e.g. name.
-        """
-        raise NotImplementedError
-
-    @property
-    def requires(self) -> tuple:
-        """The roles required to view the resource and if all roles are required.
-
-        tuple[0] = required_roles
-        tuple[1] = User must have all roles within required_roles if True
-        tuple[2] = Passed as kwargs, operating similar to User.has_access
-
-        Default is None, False, {} which means no auth.
-        """
-        return None, False, {}
-
-    @property
     def as_dict(self):
-        return self._as_dict
+        return {k: v for k, v in self.__dict__.items() if not k.startswith('_')}
 
     @classmethod
-    def _with_content(cls, **kwargs) -> dict:
-        """Retrieves data from cache or API to be used object_* classmethods
-
-        :param kwargs:
-        :return: dict
-        """
-        model = cls(**kwargs)
-        min_age = dt.now() - timedelta(minutes=model.ttl)
-        if not os.path.exists(model.file_path) or dt.fromtimestamp(os.path.getmtime(model.file_path)) < min_age:
-            return model._api_retrieve()
-        else:
-            with open(model.file_path) as config:
-                return json.loads(config.read())
-
-    @classmethod
-    def objects_keys(cls, **kwargs) -> list:
+    def objects_identifier(cls, **kwargs) -> list:
         """Returns a list of values, primarily used for passing in click.Choice(Model.objects_keys())
 
         :param kwargs:
         :return: list
         """
-        content = cls._with_content(**kwargs)
-        model = cls(**kwargs)
-        return [item.get(model.key_identifier) for item in content]
+        model = cls.objects_all(as_dict=True, **kwargs)
+        return list(model.keys())
 
     @classmethod
     def objects_all(cls, as_dict: bool = False, **kwargs):
-        """Returns all hits as a list of Model objects or dict(key_identifier=item_as_dict).
+        """Returns all hits as a list of Model objects or dict(obj_identifier=item_as_dict).
 
         :param as_dict: If true (default False), the response will be a dict instead of a list of models
         :param kwargs:
         :return:
         """
-        content = cls._with_content(**kwargs)
+        model = cls(**kwargs)
+        content = model._cache_retrieve()
+        if not content:
+            content = model.refresh()
+
         if as_dict:
-            model = cls(**kwargs)
-            return {item.get(model.key_identifier): item for item in content}
+            return {item.get(model._obj_identifier): item for item in content}
         else:
             return [cls(**{**model, **kwargs}) for model in content]
 
@@ -104,8 +74,8 @@ class Model:
         :param kwargs:
         :return: cls
         """
-        model = cls.objects_all(as_dict=True)
-        return cls(**{**model.get(key), **kwargs})
+        model_dict = cls.objects_all(as_dict=True)
+        return cls(**{**model_dict.get(key), **kwargs})
 
     def get(self, attr, default=None):
         """Perform a safe lookup on an instance of the Model with the ability to provide a default if attr not set.
@@ -119,12 +89,31 @@ class Model:
             setattr(self, attr, default)
         return val
 
-    def _api_retrieve(self):
-        """Protected method to retrieve and all Model objects. Used within Model.objects.refresh
+    def _cache_retrieve(self) -> list:
+        """Protected method to retrieve model responses from cache.
+        :return: list(requests.Response)
+        """
+        min_age = dt.now() - timedelta(minutes=self._ttl)
+        if os.path.exists(self._cache_key) and dt.fromtimestamp(os.path.getmtime(self._cache_key)) < min_age:
+            with open(self._cache_key) as config:
+                return json.loads(config.read())
+
+    def _cache_set(self, content):
+        """Called by Model.refresh if _cache_key is not None. Protected method that sets cache content.
+
+        :param content:
+        """
+        cache_key = Path(os.path.expanduser(self._cache_key))
+        os.makedirs(os.path.dirname(cache_key), exist_ok=True)
+        with open(cache_key, 'w') as config:
+            config.write(json.dumps(content, indent=2))
+
+    def _api_retrieve(self) -> list:
+        """Protected method to retrieve and all Model objects from the api.
         :return: list(requests.Response)
         """
         page = 1
-        response = self.session.get(self.url, params=dict(page=page))
+        response = self._session.get(self._url, params=dict(page=page))
         if response.status_code >= 400:
             return []
 
@@ -134,8 +123,8 @@ class Model:
 
         if len(results) < count:
             remaining_pages = math.ceil(count/len(results)) + 2  # account for 1 indexed val with page 1 complete
-            request_list = [(self.url, dict(params=dict(page=page))) for page in range(2, remaining_pages)]
-            responses = self.session.bulk_get(request_list)
+            request_list = [(self._url, dict(params=dict(page=page))) for page in range(2, remaining_pages)]
+            responses = self._session.bulk_get(request_list)
             for response in responses:
                 if response.status_code < 400:
                     content = response.json()
@@ -143,20 +132,16 @@ class Model:
 
         return results
 
-    def refresh(self):
+    def refresh(self) -> list:
         """Syncs the local file with the service
         """
-        kwargs = self.requires[2] if len(self.requires) > 2 and isinstance(self.requires[2], dict) else {}
-        if isinstance(self.session, UserSession) and not self.session.user.has_access(requires=list(self.requires[0]),
-                                                                                      match_all=bool(self.requires[1]),
-                                                                                      **kwargs):
-            return False
+        if isinstance(self._session, UserSession) and not self._session.user.has_access(requires=self._auth.requires,
+                                                                                        match_all=self._auth.match_all,
+                                                                                        **self._auth.kwargs):
+            return []
 
         content = self._api_retrieve()
-        if content and self.file_path:
-            os.makedirs(os.path.dirname(self.file_path), exist_ok=True)
-            with open(self.file_path, 'w') as config:
-                config.write(json.dumps(content, indent=4))
-
+        if content and self._cache_key:
+            self._cache_set(content)
         return content