+++ double_click/utils.py
@@ -95,6 +95,7 @@ def ensure_latest_package(package_name: str, pip_args: list = [], md_file: str =
 
     :param package_name:
     :param pip_args:
+    :param md_file:
     :return:
     """
     proc = subprocess.Popen(